+++ lesson4/drink.rb
@@ -1,0 +1,10 @@
+require "./menu"
+
+class Drink < Menu
+  attr_accessor :type
+
+  def initialize(name:, price:, type:)
+    super
+    self.type = type
+  end
+end
