+++ lesson4/food.rb
@@ -1,0 +1,4 @@
+require "./menu"
+
+class Food < Menu
+end
